refactor(proposals): extract section forms from SectionEditor into per-type components
SectionEditor.vue shrinks from 1,298 to 328 lines: the 16 inline section
forms move verbatim (byte-identical templates) to
components/BusinessProposal/admin/section-forms/, dispatched through a
sectionFormRegistry keyed by section_type — mirroring the public
sectionComponentMap. All forms mutate the parent-owned reactive form via
props (same data flow as before); Investment keeps its auto-fill watches,
ValueAddedModules its allSections computeds and FunctionalRequirements its
paste toggles inside their own components. SECTION_TYPE_OPTIONS is exported
for the upcoming add-section modal. Every data-testid and visible label
preserved; SectionEditor.test.js (173 tests) passes unchanged.
